+++ CHANGELOG.md
@@ -1,8 +1,14 @@
 # Changelog
 
+## [Unreleased]
+### Added
+- The Micro-Manager Video4Linux device adapter
+  (https://github.com/micro-manager/micro-manager/tree/mm2/DeviceAdapters/Video4Linux)
+
 ## [0.2.0]
 ### Added
 - RPi-GPIO device adapter
 - Automated deployment of the build artifacts to GitHub releases
 
+[Unreleased]: https://github.com/kmdouglass/RPi-DeviceAdapters/compare/0.2.0...HEAD
 [0.2.0]: https://github.com/kmdouglass/RPi-DeviceAdapters/releases/tag/0.2.0